+++ Assets/Scripts_2/Components/Interfaces/health_interface.cs
@@ -4,5 +4,5 @@ using System.Collections;
 
 public interface health_interface : IEventSystemHandler {
 
-    void On_Modify_Health(int _amount);
+    void On_Modify_Health(float _amount);
 }
